start_run.csh should now have the same functionality as run.csh, i.e. it listens to NEWDIR and similar tricks. In addition, if no data/proc0/var.dat exists, it runs start there too, and then runs run. Everything has been tested on my laptop, but not yet on a machine where the data are copied to a local scratch disc. Between start and run it should not move data back to the server; only for NEWDIR it should do this. So I'll now use this start_run.csh routinely instead of run.csh, even for ordinary restarts, because it allows me then to run a complete fresh start in another directory. [in run.csh I just removed some trailing empty characters which caused unnecessary differences when comparing against start_run.csh]

+++ bin/start_run.csh
@@ -1,5 +1,5 @@
 #!/bin/csh
-# CVS: $Id: start_run.csh,v 1.41 2006-12-06 19:53:17 wlyra Exp $
+# CVS: $Id: start_run.csh,v 1.42 2007-03-06 17:35:25 brandenb Exp $
 
 #                       start_run.csh
 #                      ---------------
@@ -9,6 +9,7 @@
 #
 # Run this script with csh:
 #PBS -S /bin/csh
+#PBS -r n
 #$ -S /bin/csh
 #@$-s /bin/csh
 #
@@ -25,11 +26,79 @@ if ($?QSUB_WORKDIR) then
   cd $QSUB_WORKDIR
 endif
 
-# ---------------------------------------------------------------------- #
+# Work in submit directory (IBM Loadleveler):
+if ($?LOADL_STEP_INITDIR) then
+  cd $LOADL_STEP_INITDIR
+endif
+
+# ====================================================================== #
+# Starting points when rerunning in new directory.
+# This needs to come *before* "source getconf.csh", because
+#  (i) it checks for LOCK file, and
+# (ii) it sets $datadir/directory_snap
+newdir:
 
 # Common setup for start.csh, run.csh, start_run.csh:
 # Determine whether this is MPI, how many CPUS etc.
 source getconf.csh
+
+# ====================================================================== #
+#  Assume that if data/proc0/var.dat exists, we are not supposed
+#  to overwrite it, and should try to restart from the existing
+#  snapshot instead. In that case, if $local_disc was set in getconf.csh,
+#  we need to copy these data first to the locak scratch disc, and
+#  then restart, i.e. jump to rerun.
+#
+if (-e "$datadir"/proc0/var.dat) then
+#
+#  If necessary, distribute var.dat from the server to the various nodes
+#  Don't indent these lines so that it is easier to vimdiff against run.csh
+#
+if ($local_disc) then
+  if ($one_local_disc) then     # one common local disc
+    foreach node ($nodelist)
+      foreach d (`cd $datadir; ls -d proc* allprocs`)
+        if (-e $datadir/$d/var.dat) $SCP $datadir/$d/var.dat ${node}:$SCRATCH_DIR/$d/
+        if (-e $datadir/$d/global.dat) $SCP $datadir/$d/global.dat ${node}:$SCRATCH_DIR/$d/
+        if ($lparticles) $SCP $datadir/$d/pvar.dat ${node}:$SCRATCH_DIR/$d/
+        $SCP $datadir/$d/timeavg.dat ${node}:$SCRATCH_DIR/$d/
+      end
+      if (-e $datadir/allprocs/dxyz.dat) $SCP $datadir/allprocs/dxyz.dat ${node}:$SCRATCH_DIR/allprocs
+    end
+  else # one local disc per MPI process (Horseshoe, etc);
+       # still doesn't cover Copson
+    set i = -1
+    foreach node ($nodelist)
+      set i=`expr $i + 1`
+      echo "i = $i"
+      set j = 0
+      while ($j != $nprocpernode)
+        set k = `expr $nprocpernode \* $i + $j`
+        if ($?notserial_procN) set k = `expr $i + $nnodes \* $j`
+        $SCP $datadir/proc$k/var.dat ${node}:$SCRATCH_DIR/proc$k/
+        if (-e $datadir/proc$k/global.dat) then
+          $SCP $datadir/proc$k/global.dat ${node}:$SCRATCH_DIR/proc$k/
+        endif
+        if ($lparticles) then
+          $SCP $datadir/proc$k/pvar.dat ${node}:$SCRATCH_DIR/proc$k/
+        endif
+        echo "$SCP $datadir/proc$k/var.dat ${node}:$SCRATCH_DIR/proc$k/"
+        if (-e $datadir/proc$k/timeavg.dat) then
+          $SCP $datadir/proc$k/timeavg.dat ${node}:$SCRATCH_DIR/proc$k/
+        endif
+        set j=`expr $j + 1`
+      end
+      if (-e $datadir/allprocs/dxyz.dat) then
+        $SCP $datadir/allprocs/dxyz.dat ${node}:$SCRATCH_DIR/allprocs/
+      endif
+    end
+  endif
+endif
+#
+# goto rerun instead of producing a new initial condition
+#
+goto rerun
+endif
 
 #
 #  If we don't have a data subdirectory: stop here (it is too easy to
@@ -43,9 +112,34 @@ if (! -d "$datadir") then
   echo ">>  but that will most likely end up on your NFS file system and be"
   echo ">>  slow"
   echo
-  rm -f LOCK
-  exit 0
+  goto checknewdir
 endif
+
+#
+#  Execute some script or command specified by the user.
+#  E.g. run src-to-data to save disk space in /home by moving src/ to
+#  data/ and linking
+#
+if ($?PENCIL_START1_CMD) then
+  echo "Running $PENCIL_START1_CMD"
+  $PENCIL_START1_CMD
+endif
+
+# ---------------------------------------------------------------------- #
+
+#  For testing backwards compatibility, do not excecute start.x, an do not
+#  delete existing var.dat (etc.) files.
+#  Rename time_series.dat, so run.x can write a fresh one to compare
+#  against.
+#
+if (-e NOSTART) then
+  echo "Found NOSTART file. Won't run start.x"
+  if (-e $datadir/time_series.dat) \
+      mv $datadir/time_series.dat $datadir/time_series.`timestr`
+  exit
+endif
+
+# ---------------------------------------------------------------------- #
 
 # Create list of subdirectories
 # If the file NOERASE exists, the old directories are not erased
@@ -74,15 +168,13 @@ if (! -e NOERASE) then
       mv $datadir/time_series.dat $datadir/time_series.`timestr`
   rm -f $datadir/*.dat $datadir/*.nml $datadir/param*.pro $datadir/index*.pro \
         $datadir/averages/* >& /dev/null
-  rm -f $datadir/move-me.list $datadir/moved-files.list >& /dev/null
+  if ($lcopysnapshots_exp) rm -f $datadir/move-me.list $datadir/moved-files.list >& /dev/null
 endif
 
-# If local disk is used, copy executable to $SCRATCH_DIR of master node
+# If local_binary is used, copy executable to $SCRATCH_DIR of master node
 if ($local_binary) then
+  echo "Copying start.x to $SCRATCH_DIR"
   cp src/start.x $SCRATCH_DIR
-endif
-# Copy output from `top' on run host to a file we can read from login server
-if ($remote_top) then
   remote-top >& remote-top.log &
 endif
 
@@ -95,27 +187,31 @@ if ($start_status) exit $start_status	# something went wrong
 echo ""
 date
 
+# Not sure it makes any sense to continue after mpirun had an error:
+if ($status) then
+  echo "Error status $status found -- aborting"
+  exit $start_status
+endif
+
+# ---------------------------------------------------------------------- #
+rerun:
+
 # Clean up control and data files
-rm -f STOP RELOAD fort.20
+# NB. Don't remove NEWDIR it may have been put there on purpose so as
+#     to catch a crash and run something else instead.
+rm -f STOP RELOAD RERUN fort.20
 
+# On machines with local scratch directory, initialize automatic
+# background copying of snapshots back to the data directory.
+# Also, if necessary copy executable to $SCRATCH_DIR of master node
+# and start top command on all procs.
 if ($local_disc) then
-  # We still need to copy (at least one of) the var.dat files back, so
-  # the background process copy-snapshots will know how large the snapshots
-  # ought to be. Certainly far from elegant..
-  $copysnapshots -v var.dat     >&  copy-snapshots.log
-  if ($lparticles) $copysnapshots -v pvar.dat >>& copy-snapshots.log
-  $copysnapshots -v timeavg.dat >>& copy-snapshots.log
-
-  # On machines with local scratch directory, initialize automatic
-  # background copying of snapshots back to the data directory.
-  # Also, if necessary copy executable to $SCRATCH_DIR of master node
-  # and start top command on all procs.
   echo "Use local scratch disk"
-  $copysnapshots -v >>& copy-snapshots.log &
+  $copysnapshots -v >&! copy-snapshots.log &
 endif
 # Copy output from `top' on run host to a file we can read from login server
 if ($remote_top) then
-  remote-top >& remote-top.log &
+  remote-top >&! remote-top.log &
 endif
 if ($local_binary) then
   echo "ls src/run.x $SCRATCH_DIR before copying:"
@@ -125,6 +221,16 @@ if ($local_binary) then
   ls -lt src/run.x $SCRATCH_DIR
 endif
 
+# Write $PBS_JOBID or $LOADL_STEP_ID to file
+# (important when run is migrated within the same job)
+if ($?PBS_JOBID) then
+  echo $PBS_JOBID "  RUN STARTED on "$PBS_O_QUEUE `date` \
+    >> $datadir/jobid.dat
+endif
+if ($?LOADL_STEP_ID) then
+  echo $LOADL_STEP_ID "  RUN STARTED on "$LOADL_STEP_CLASS `date` \
+    >> $datadir/jobid.dat
+endif
 
 # Run run.x
 date
@@ -133,6 +239,30 @@ echo $mpirun $mpirunops $npops $mpirunops2 $run_x $x_ops >! run_command.log
 time $mpirun $mpirunops $npops $mpirunops2 $run_x $x_ops
 set run_status=$status		# save for exit
 date
+
+# Create symlinks for deprecated slices
+pc_deprecated_slice_links
+
+# Write $PBS_JOBID to file (important when run is migrated within the same job)
+if ($?PBS_JOBID) then
+  echo $PBS_JOBID " RUN FINISHED on "$PBS_O_QUEUE `date` >> $datadir/jobid.dat
+endif
+
+# look for RERUN file 
+# With this method one can only reload a new executable.
+# One cannot change directory, nor are the var.dat files returned to server.
+# See the NEWDIR method below for more options.
+if (-e "RERUN") then
+  rm -f RERUN
+  echo
+  echo "======================================================================="
+  echo "Rerunning in the *same* directory; current run status: $run_status"
+  echo "We are *still* in: " `pwd`
+  echo "======================================================================="
+  echo
+  goto rerun
+endif
+# ---------------------------------------------------------------------- #
 
 # On machines with local scratch disc, copy var.dat back to the data
 # directory
@@ -162,24 +292,48 @@ if ($local_disc) then
 endif
 echo "Done"
 
+checknewdir:
+
+# look for NEWDIR file 
+# if NEWDIR contains a directory name, then continue run in that directory
+if (-e "NEWDIR") then 
+  if (-s "NEWDIR") then
+    # Remove LOCK file before going to other directory
+    if (-e "LOCK") rm -f LOCK
+    set olddir=$cwd
+    cd `cat NEWDIR`
+    rm $olddir/NEWDIR
+    (echo "stopped run:"; date; echo "new run directory:"; echo $cwd; echo "")\
+       >> $olddir/$datadir/directory_change.log
+    (date; echo "original run script is in:"; echo $olddir; echo "")\
+       >> $datadir/directory_change.log
+    echo
+    echo "====================================================================="
+    echo "Rerunning in new directory; current run status: $run_status"
+    echo "We are now in: " `pwd`
+    echo "====================================================================="
+    echo
+    goto newdir
+  else
+    rm -f NEWDIR
+    if (-e "LOCK") rm -f LOCK
+    echo
+    echo "====================================================================="
+    echo "Rerunning in the *same* directory; current run status: $run_status"
+    echo "We are *still* in: " `pwd`
+    echo "====================================================================="
+    echo
+    echo "Rerunning; current run status: $run_status"
+    goto newdir
+  endif
+endif
+# ====================================================================== #
+
 # Shut down lam if we have started it
 if ($booted_lam) lamhalt
 
 # remove LOCK file
 if (-e "LOCK") rm -f LOCK
-
-# look for RERUN file 
-if (-e "RERUN") then 
-  if (-s "RERUN") then
-    cd `cat RERUN`
-  endif
-
-  rm -f RERUN
-
-  echo "Rerunning; current run status: $run_status"
-  #  goto rerun
-  ./run.csh
-endif  
 
 exit $run_status		# propagate status of mpirun
 
@@ -193,3 +347,4 @@ exit $run_status		# propagate status of mpirun
 # qsub -l nodes=4:ppn=1,mem=500mb,cput=24:00:00 -q p-long start_run.csh
 # qsub -l ncpus=4,mem=1gb,cput=100:00:00 -q parallel start_run.csh
 # qsub -l nodes=128,mem=64gb,walltime=1:00:00 -q workq start_run.csh
+# eval `env-setup lam`; qsub -v PATH -pe lam 8 -j y -o run.log run.csh
